@@ -1,11 +1,11 @@
-import { Navigation } from './routes/Navigation';
+import Navigation from './routes/Navigation';
 
 function App() {
-  return (
-    <>
-      <Navigation />
-    </>
-  );
+    return (
+        <>
+            <Navigation />
+        </>
+    );
 }
 
 export default App;
